Partially fix #69: + If password is empty, ignore silently. No error, no update. + "hostname" and "regenerate keys" as advanced config parameters.

+++ lib/security_config_handler.py
@@ -46,11 +46,6 @@ class SecurityConfigHandler(ZynthianConfigHandler):
 	def get(self, errors=None):
 		#Get Hostname
 		config=OrderedDict([
-			['HOSTNAME', {
-				'type': 'text',
-				'title': 'Hostname',
-				'value': SecurityConfigHandler.get_host_name()
-			}],
 			['PASSWORD', {
 				'type': 'password',
 				'title': 'Password',
@@ -61,12 +56,19 @@ class SecurityConfigHandler(ZynthianConfigHandler):
 				'title': 'Repeat password',
 				'value': '*'
 			}],
+			['HOSTNAME', {
+				'type': 'text',
+				'title': 'Hostname',
+				'value': SecurityConfigHandler.get_host_name(),
+				'advanced': True
+			}],
 			['REGENERATE_KEYS', {
 				'type': 'button',
 				'title': 'Regenerate Keys',
 				'script_file': 'regenerate_keys.js',
 				'button_type': 'button',
-				'class': 'btn-warning',
+				'class': 'btn-warning btn-block',
+				'advanced': True
 			}],
 			['_command', {
 				'type': 'hidden',
@@ -94,12 +96,17 @@ class SecurityConfigHandler(ZynthianConfigHandler):
 
 	def update_system_config(self, config):
 		#Update Password
-		if len(config['PASSWORD'][0])<6:
-			return { 'PASSWORD': "Password must have at least 6 characters" }
-		if config['PASSWORD'][0]!=config['REPEAT_PASSWORD'][0]:
-			return { 'REPEAT_PASSWORD': "Passwords does not match!" }
-		esc_passwd=config['PASSWORD'][0].replace('"','\\"')
-		check_output("echo -e \"root:{}\" | chpasswd".format(esc_passwd), shell=True, executable="/bin/bash")
+		if len(config['PASSWORD'][0])>0:
+			if len(config['PASSWORD'][0])<6:
+				return { 'PASSWORD': "Password must have at least 6 characters" }
+			if config['PASSWORD'][0]!=config['REPEAT_PASSWORD'][0]:
+				return { 'REPEAT_PASSWORD': "Passwords does not match!" }
+			try:
+				esc_passwd=config['PASSWORD'][0].replace('"','\\"')
+				check_output("echo -e \"root:{}\" | chpasswd".format(esc_passwd), shell=True, executable="/bin/bash")
+			except Exception as e:
+				return { 'REPEAT_PASSWORD': "Can't change password ({})".format(e) }
+			
 
 		#Update Hostname
 		newHostname = config['HOSTNAME'][0]
@@ -108,15 +115,17 @@ class SecurityConfigHandler(ZynthianConfigHandler):
 			previousHostname=f.readline()
 			f.close()
 
-		with open("/etc/hostname",'w') as f:
-			f.write(newHostname)
-			f.close()
+		if previousHostname!=newHostname:
+			with open("/etc/hostname",'w') as f:
+				f.write(newHostname)
+				f.close()
 
-		with open("/etc/hosts", "r+") as f:
-			contents = f.read()
-			contents = contents.replace(previousHostname, newHostname)
-			contents = contents.replace("zynthian", newHostname) # for the ppl that have already changed their hostname
-			f.seek(0)
-			f.truncate()
-			f.write(contents)
-			f.close()
+			with open("/etc/hosts", "r+") as f:
+				contents = f.read()
+				contents = contents.replace(previousHostname, newHostname)
+				contents = contents.replace("zynthian", newHostname) # for the ppl that have already changed their hostname
+				f.seek(0)
+				f.truncate()
+				f.write(contents)
+				f.close()
+
